Add legacy events to dataLayer
