+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Legacy events to dataLayer:
+  - `homeView`
+  - `categoryView`
+  - `departmentView`
+  - `internalSiteSearchView`
+  - `productView`
+  - `otherView`
 
 ## [2.8.1] - 2021-01-12
 ### Fixed
